# app/microservice/public-swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "RW Metadata API",
        "description": "Resource Watch - Metadata API",
        "version": "1.0.0"
    },
    "host": "example.vizzuality.com",
    "schemes": [
        "https",
        "http"
    ],
    "produces": [
        "application/vnd.api+json"
    ],
    "paths": {
        "/metadata": {
            "post": {
                "description": "Get metadatas of a datasets and application",
                "operationId": "findByIds",
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "METADATA"
                ],
                "parameters": [
                    {
                        "name": "metadata",
                        "in": "body",
                        "description": "Ids of datasets",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/MetadataFind"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadatas obtained",
                        "schema": {
                            "$ref": "#/definitions/MetadataAPI"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            }
        },
        "/metadata/{dataset}": {
            "get": {
                "description": "Returns metadatas filtered by dataset",
                "operationId": "queryByDataset",
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "METADATA"
                ],
                "parameters": [
                    {
                        "name": "dataset",
                        "in": "path",
                        "description": "ID of dataset",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadatas obtained",
                        "schema": {
                            "$ref": "#/definitions/MetadataAPI"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove all metadata of a dataset",
                "operationId": "deleteByDataset",
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "METADATA"
                ],
                "parameters": [
                    {
                        "name": "dataset",
                        "in": "path",
                        "description": "ID of dataset",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadatas removed",
                        "schema": {
                            "$ref": "#/definitions/MetadataAPI"
                        }
                    },
                    "404": {
                        "description": "Metadata not found",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            }
        },
        "/metadata/{dataset}/{application}": {
            "get": {
                "description": "Returns metadatas filtered by dataset",
                "operationId": "queryByDatasetAndApplication",
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "METADATA"
                ],
                "parameters": [
                    {
                        "name": "dataset",
                        "in": "path",
                        "description": "ID of dataset",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "application",
                        "in": "path",
                        "description": "Slug of application (gfw,rw,prep)",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadatas obtained",
                        "schema": {
                            "$ref": "#/definitions/MetadataAPI"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create metadata",
                "description": "Create Metadata",
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "METADATA"
                ],
                "operationId": "create",
                "parameters": [
                    {
                        "name": "dataset",
                        "in": "path",
                        "description": "ID of dataset",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "application",
                        "in": "path",
                        "description": "Slug of application (gfw,rw,prep)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "metadata",
                        "in": "body",
                        "description": "Metadata to save. Free object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Metadata"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Metadata created",
                        "schema": {
                            "$ref": "#/definitions/MetadataAPI"
                        }
                    },
                    "400": {
                        "description": "Validate error or metadata duplicated",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove metadata of a dataset and an application",
                "operationId": "deleteByDatasetAndApplication",
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "METADATA"
                ],
                "parameters": [
                    {
                        "name": "dataset",
                        "in": "path",
                        "description": "ID of dataset",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "application",
                        "in": "path",
                        "description": "Slug of application (gfw,rw,prep)",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Metadatas removed",
                        "schema": {
                            "$ref": "#/definitions/MetadataAPI"
                        }
                    },
                    "404": {
                        "description": "Metadata not found",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            },
            "patch": {
                "summary": "Update metadata",
                "description": "Update Metadata",
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "METADATA"
                ],
                "operationId": "update",
                "parameters": [
                    {
                        "name": "dataset",
                        "in": "path",
                        "description": "ID of dataset",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "application",
                        "in": "path",
                        "description": "Slug of application (gfw,rw,prep)",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "metadata",
                        "in": "body",
                        "description": "Metadata to save. Free object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Metadata"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Metadata updated",
                        "schema": {
                            "$ref": "#/definitions/MetadataAPI"
                        }
                    },
                    "400": {
                        "description": "Validate error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Errors": {
            "type": "object",
            "description": "Errors",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Error"
                    }
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32",
                    "description": "A unique identifier for this particular occurrence of the problem."
                },
                "links": {
                    "type": "object",
                    "description": "A links object",
                    "properties": {
                        "about": {
                            "type": "string",
                            "description": "A link that leads to further details about this particular occurrence of the problem."
                        }
                    }
                },
                "status": {
                    "type": "string",
                    "description": "The HTTP status code applicable to this problem, expressed as a string value"
                },
                "code": {
                    "type": "string",
                    "description": "An application-specific error code, expressed as a string value"
                },
                "title": {
                    "type": "string",
                    "description": "A short, human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the problem, except for purposes of localization."
                },
                "detail": {
                    "type": "string",
                    "description": "A human-readable explanation specific to this occurrence of the problem. Like title, this field's value can be localized"
                },
                "source": {
                    "type": "object",
                    "description": "An object containing references to the source of the error, optionally including any of the following members",
                    "properties": {
                        "pointer": {
                            "type": "string",
                            "description": "A JSON Pointer [RFC6901] to the associated entity in the request document"
                        },
                        "parameter": {
                            "type": "string",
                            "description": "A string indicating which URI query parameter caused the error."
                        }
                    }
                },
                "meta": {
                    "type": "object",
                    "description": "A meta object containing non-standard meta-information about the error."
                }
            }
        },
        "MetadataFind": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "description": "ids of datasets",
                    "items": {
                        "type": "string"
                    }
                },
                "app": {
                    "type": "string",
                    "description": "Name of the application (gfw, prep, etc)"
                }
            }
        },
        "MetadataAPI": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Metadata"
                    }
                }
            }
        },
        "Metadata": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string",
                    "description": "Resource type"
                },
                "id": {
                    "type": "string",
                    "description": "Resource id"
                },
                "attributes": {
                    "properties": {
                        "schema": {
                            "$ref": "#/definitions/MetadataBodyResponse"
                        }
                    }
                }
            }
        },
        "MetadataBodyResponse": {
            "type": "object",
            "properties": {
                "dataset": {
                    "type": "string",
                    "description": "Id of dataset"
                },
                "application": {
                    "type": "string",
                    "description": "Slug of an application"
                },
                "info": {
                    "type": "object",
                    "description": "Free object with the info of metadata"
                }
            }
        }
    }
}